[MOV] im_livechat: move AbstractMessage outside of public_livechat.js
Task-2212347
